Patch for race on HTTP-100 expected continue
This patch is a first-pass attempt to fix the issue described by
golang/go#34902, where the use of 100-continue in HTTP headers received
by a net/http/Server can lead to a data race involving the connection's
buffered writer and the expectContinueReader.sawEOF field.

This patch is not activated unless the process's environment variables
include NET_HTTP_SERVER_SYNC_BUFIO_WRITER=1.

This patch also includes a test to validate the race condition no longer
occurs. To execute the test to validate the existence of the race
condition please use:

        GORACE="halt_on_error=1" go test -race \
          -v -run TestExpectedContinueRace net/http

To execute the test to validate the race condition no longer exists once
the patch is applied, please use:

        NET_HTTP_SERVER_SYNC_BUFIO_WRITER=1 \
          GORACE="halt_on_error=1" go test -race \
          -v -run TestExpectedContinueRace net/http

The test allows five minutes for the race to occur, after which the
absence of the race is deemed as evidence it no longer occurs.